pyproject.toml
[tool.poetry]
name = "smart-test-framework"
version = "0.1.0"
description = "智能分层测试框架"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
playwright = "^1.40.0"
requests = "^2.31.0"
pytest = "^7.4.3"
pytest-xdist = "^3.3.1"
allure-pytest = "^2.13.2"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
faker = "^22.0.0"
opencv-python = "^4.8.1"
openai = "^1.3.0"
pytest-playwright = "^1.40.0"
loguru = "^0.7.2"
pytest-html = "^4.1.1"
pytest-rerunfailures = "^12.0"
pytest-timeout = "^2.2.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3 